feat(android.FavoriteConfirmation): Don't show dialog if only one direction
